freeflow/template/requirements.txt
apache-airflow[crypto]==1.9.0

